// backend/src/api/whatsapp/enums/whatsapp.enum.ts
export enum WhatsAppDirection {
  INBOUND = 'INBOUND',
  OUTBOUND = 'OUTBOUND',
}
export enum WhatsAppMessageType {
  TEXT = 'TEXT',
  IMAGE = 'IMAGE',
  AUDIO = 'AUDIO',
  STATUS = 'STATUS',
}
export enum WhatsAppMessageStatus {
  RECEIVED = 'RECEIVED',
  PROCESSING = 'PROCESSING',
  SENT = 'SENT',
  DELIVERED = 'DELIVERED',
  READ = 'READ',
  FAILED = 'FAILED',
  IGNORED = 'IGNORED',
}